Moves test utils out of common, and into test-specific common.
